[OPIK-1039] refactoring
