+++ apps/opik-frontend/src/components/pages/ExperimentsPage/charts/ExperimentsChartsWrapper.tsx
@@ -10,20 +10,19 @@ import { formatDate } from "@/lib/date";
 import ExperimentChartContainer, {
   ChartData,
 } from "@/components/pages/ExperimentsPage/charts/ExperimentChartContainer";
-import { Spinner } from "@/components/ui/spinner";
+import { Dataset } from "@/types/datasets";
 
 export type ExperimentsChartsWrapperProps = {
   experiments: GroupedExperiment[];
-  datasetsIds: string[];
+  datasetsData: Dataset[];
 };
 
 const ExperimentsChartsWrapper: React.FC<ExperimentsChartsWrapperProps> = ({
   experiments,
-  datasetsIds,
+  datasetsData,
 }) => {
   const chartsData = useMemo(() => {
     const groupsMap: Record<string, ChartData> = {};
-    let index = 0;
 
     experiments.forEach((experiment) => {
       if (experiment.virtual_dataset_id !== DELETED_DATASET_ID) {
@@ -32,9 +31,7 @@ const ExperimentsChartsWrapper: React.FC<ExperimentsChartsWrapperProps> = ({
             dataset: experiment.dataset,
             data: [],
             lines: [],
-            index,
           };
-          index += 1;
         }
 
         groupsMap[experiment.virtual_dataset_id].data.unshift({
@@ -60,9 +57,9 @@ const ExperimentsChartsWrapper: React.FC<ExperimentsChartsWrapperProps> = ({
   }, [experiments]);
 
   const chartClassName =
-    datasetsIds.length === 1
+    datasetsData.length === 1
       ? "w-full"
-      : datasetsIds.length === 2
+      : datasetsData.length === 2
         ? "basis-1/2"
         : "basis-[520px]";
 
@@ -70,33 +67,18 @@ const ExperimentsChartsWrapper: React.FC<ExperimentsChartsWrapperProps> = ({
     <div
       className={cn(
         "flex items-center gap-4 overflow-y-auto",
-        datasetsIds.length > 0 && "mb-4",
+        datasetsData.length > 0 && "mb-4",
       )}
     >
-      {datasetsIds.map((datasetId) => {
-        const chartData = chartsData[datasetId];
-
-        if (!chartData) {
-          return (
-            <div
-              key={datasetId}
-              className={cn(
-                "flex h-52 w-full items-center justify-center rounded-lg border bg-card shadow-sm",
-                chartClassName,
-              )}
-            >
-              <Spinner />
-            </div>
-          );
-        }
-        return (
-          <ExperimentChartContainer
-            key={datasetId}
-            chartData={chartData}
-            className={chartClassName}
-          />
-        );
-      })}
+      {datasetsData.map((dataset) => (
+        <ExperimentChartContainer
+          key={dataset.id}
+          chartId={dataset.id}
+          chartData={chartsData[dataset.id]}
+          dataset={dataset}
+          className={chartClassName}
+        />
+      ))}
     </div>
   );
 };
